ZeroBezel code review suggestions implemented: 1. Modularised the existing file structure 2. Created generic methods for common code
