+++ CTNotificationContent/Templates/Assets.xcassets/AccentColor.colorset/Contents.json
