+++ CTNotificationContent/Templates/Assets.xcassets/AppIcon.appiconset/Contents.json
