+++ CTNotificationContent/Templates/Assets.xcassets/Contents.json
