+++ CTNotificationContent/Templates/Carousel/Model/CarouselProperties.swift
@@ -1,0 +1,22 @@
+//
+//  CarouselProperties.swift
+//  CleverTap-iOS-SDK
+//
+//  Created by Aishwarya Nanna on 26/08/22.
+//
+
+import Foundation
+
+struct CarouselProperties: Decodable {
+    let pt_title: String?
+    let pt_msg: String?
+    let pt_msg_summary: String?
+    let pt_dl1: String?
+    let pt_big_img: String?
+    let pt_img1: String?
+    let pt_img2: String?
+    let pt_img3: String?
+    let pt_bg: String?
+    let pt_title_clr: String?
+    let pt_msg_clr: String?
+}
