+++ CTNotificationContent/Templates/SingleMedia/Controller/CTSingleMediaController.swift
