+++ CTNotificationContent/Templates/Utility/GlobalConstants.swift
