+++ CTNotificationContent/Templates/ZeroBezel/Model/ZeroBezelProperties.swift
@@ -1,0 +1,20 @@
+//
+//  ZeroBezelProperties.swift
+//  CleverTap-iOS-SDK
+//
+//  Created by Aishwarya Nanna on 26/08/22.
+//
+
+import Foundation
+
+struct ZeroBezelProperties: Decodable {
+    let pt_title: String?
+    let pt_msg: String?
+    let pt_msg_summary: String?
+    let pt_subtitle: String?
+    let pt_big_img: String?
+    let pt_dl1: String?
+    let pt_title_clr: String?
+    let pt_msg_clr: String?
+    let pt_small_icon_clr: String?
+}
